Move type def generator to its own file.
